Fix Error Message And Cause of Error
Per Bryan's request, I've updated the generic error message on the Data Portal site so that it doesn't mention contacting the Environmental Data Initiative, and provides a contact email in the case of repeated errors.

This error was most often triggered by users submitting an empty search, so I added some Javascript to disable the search button on the home page until the search term is at least three letters (an arbitrary value that felt about right).