Fix Background Color on Highlighted Tab
There is a section on this page that shows provenance metadata in both HTML and XML, with each being on a tab. Using generic Bootstrap nav items resulted in the tabs picking up styling from the main menu at the top of the page - including wholly incorrect and ugly styling on hover. Wrapped the tabs in a <div> and applied a more specific override to prevent this.