Remove GIOS Specific References
Removed some GIOS-specific text from a few places:

* The 'Contact' page
* The 'About' page
* The DataPortalServlet Java file itself

While neither the 'Contact' or 'About' pages are linked to, but they do exist so I figured I'd make sure they never specifically said 'GIOS'. The DataPortalServlet change was to take 'GIOS' out of the function that returns the title of any page. We don't actually use this, as a fixed title is set in our common header. Still, wanted to remove it just in case.